requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
distro==1.8.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.12.1
gotrue==2.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.3.8
packaging==23.2
pandas==2.1.4
postgrest==0.13.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.2.1
supafunc==0.3.1
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.36.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
websockets==11.0.3

